move methods can can be static to a static helper class
