+++ src/main/java/xfacthd/framedblocks/common/compat/jei/camo/CamoRecipeManagerPlugin.java
@@ -31,7 +31,7 @@ public final class CamoRecipeManagerPlugin implements ISimpleRecipeManagerPlugin
     public boolean isHandledInput(ITypedIngredient<?> ingredient)
     {
         ItemStack itemStack = ingredient.getItemStack().orElse(ItemStack.EMPTY);
-        if (camoCraftingHelper.getFramedBlock(itemStack) != null)
+        if (CamoItemStackHelper.getFramedBlock(itemStack) != null)
         {
             return true;
         }
@@ -39,17 +39,17 @@ public final class CamoRecipeManagerPlugin implements ISimpleRecipeManagerPlugin
         {
             return true;
         }
-        return CamoCraftingHelper.getCamoContainerFactory(itemStack) != null;
+        return CamoItemStackHelper.getCamoContainerFactory(itemStack) != null;
     }
 
     @Override
     public boolean isHandledOutput(ITypedIngredient<?> ingredient)
     {
         ItemStack itemStack = ingredient.getItemStack().orElse(ItemStack.EMPTY);
-        IFramedBlock framedBlock = camoCraftingHelper.getFramedBlock(itemStack);
+        IFramedBlock framedBlock = CamoItemStackHelper.getFramedBlock(itemStack);
         if (framedBlock != null)
         {
-            long camoCount = camoCraftingHelper.dropCamo(itemStack).count();
+            long camoCount = CamoItemStackHelper.dropCamo(itemStack).count();
             return camoCount > 0 && camoCount <= 2;
         }
         return false;
@@ -59,7 +59,7 @@ public final class CamoRecipeManagerPlugin implements ISimpleRecipeManagerPlugin
     public List<RecipeHolder<CraftingRecipe>> getRecipesForInput(ITypedIngredient<?> ingredient)
     {
         ItemStack itemStack = ingredient.getItemStack().orElse(ItemStack.EMPTY);
-        IFramedBlock framedBlock = camoCraftingHelper.getFramedBlock(itemStack);
+        IFramedBlock framedBlock = CamoItemStackHelper.getFramedBlock(itemStack);
         if (framedBlock != null)
         {
             return createRepresentativeRecipes(
@@ -77,7 +77,7 @@ public final class CamoRecipeManagerPlugin implements ISimpleRecipeManagerPlugin
             );
         }
 
-        CamoContainerFactory<?> containerFactory = CamoCraftingHelper.getCamoContainerFactory(itemStack);
+        CamoContainerFactory<?> containerFactory = CamoItemStackHelper.getCamoContainerFactory(itemStack);
         if (containerFactory != null)
         {
             return createRepresentativeRecipes(
@@ -95,11 +95,11 @@ public final class CamoRecipeManagerPlugin implements ISimpleRecipeManagerPlugin
     {
         ItemStack itemStack = ingredient.getItemStack().orElse(ItemStack.EMPTY);
 
-        IFramedBlock framedBlock = camoCraftingHelper.getFramedBlock(itemStack);
+        IFramedBlock framedBlock = CamoItemStackHelper.getFramedBlock(itemStack);
         if (framedBlock != null)
         {
             ItemStack plainFrame = new ItemStack(itemStack.getItem());
-            List<ItemStack> camoBlocks = camoCraftingHelper.dropCamo(itemStack).toList();
+            List<ItemStack> camoBlocks = CamoItemStackHelper.dropCamo(itemStack).toList();
             int camoCount = camoBlocks.size();
             if (camoCount == 1)
             {
@@ -152,7 +152,7 @@ public final class CamoRecipeManagerPlugin implements ISimpleRecipeManagerPlugin
             );
             recipes.add(singleRecipe);
 
-            if (camoCraftingHelper.isDoubleFramedBlock(framedBlock))
+            if (CamoItemStackHelper.isDoubleFramedBlock(framedBlock))
             {
                 // calculating the correct outputs here is impossible, leave them to be generated by
                 // CamoCraftingRecipeExtension.onDisplayedIngredientsUpdate
